# docs/swagger.yaml
basePath: /
definitions:
...
  dto.Order:
    properties:
      customerId:
        type: integer
      orderProducts:
        items:
          $ref: '#/definitions/dto.OrderProduct'
        type: array
      orderStatus:
        type: string
      paymentId:
        type: integer
      ticketNumber:
        type: integer
      totalPrice:
        type: number
    required:
    - orderProducts
    - paymentId
    - totalPrice
    type: object
  dto.OrderProduct:
    properties:
      productId:
        type: integer
      productPrice:
        type: number
    required:
    - productId
    - productPrice
    type: object
  dto.OrderProductResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.OrderResponse:
    properties:
      customerName:
        type: string
      deliveredAt:
        type: string
      doneAt:
        type: string
      notDeliveredAt:
        type: string
      orderDate:
        type: string
      orderId:
        type: integer
      orderProducts:
        items:
          $ref: '#/definitions/dto.OrderProductResponse'
        type: array
      orderStatus:
        type: string
      preparingAt:
        type: string
      ticketNumber:
        type: integer
    type: object
...
host: localshot:3210
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API for the Tech1 Fiap Project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tech1 API Docs
  version: "1.0"
paths:
...
  /api/orders:
    post:
      consumes:
      - application/json
      description: |-
        Create new order. To make an order the payment needs to be completed
        A new Ticket will be generated by the Order Date starting from 1
        In the next day the Ticket number will starts from 1 and so on
      parameters:
      - description: order
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderResponse'
        "400":
          description: Order has required fields
      summary: Create new order
      tags:
      - Order
  /api/orders/{id}:
    get:
      consumes:
      - application/json
      description: Get an order by Id
      parameters:
      - description: "12"
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderResponse'
        "400":
          description: Order has required fields
      summary: Get order by Id
      tags:
      - Order
...
  /api/orders/{id}/done:
    put:
      consumes:
      - application/json
      description: Update an order. This service wil be used by the kitchen to notify
        a customer and the waiter that the order is done
      parameters:
      - description: "12"
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Order not found
        "428":
          description: 'Precondition failed: Need to be with status Preparando'
      summary: Update an order to DONE
      tags:
      - Order
...
  /api/orders/{id}/preparing:
    put:
      consumes:
      - application/json
      description: Update an order. This service wil be used by the kitchen to notify
        a customer that the order is being prepared
      parameters:
      - description: "12"
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Order not found
        "428":
          description: 'Precondition failed: Need to be with status Criado'
      summary: Update an order to PREPARING
      tags:
      - Order
...
swagger: "2.0"